client-core: move init helpers to module
