+++ clients/client-core/src/init/mod.rs
@@ -3,27 +3,22 @@
 
 //! Collection of initialization steps used by client implementations
 
-use std::{fmt::Display, sync::Arc, time::Duration};
+use std::fmt::Display;
+
+use nymsphinx::addressing::{clients::Recipient, nodes::NodeIdentity};
+use serde::Serialize;
+use tap::TapFallible;
 
 use config::NymConfig;
 use crypto::asymmetric::{encryption, identity};
-use gateway_client::GatewayClient;
-use gateway_requests::registration::handshake::SharedKeys;
-use nymsphinx::addressing::clients::Recipient;
-use nymsphinx::addressing::nodes::NodeIdentity;
-use rand::rngs::OsRng;
-use rand::seq::SliceRandom;
-use rand::thread_rng;
-use serde::Serialize;
-use tap::TapFallible;
-use topology::{filter::VersionFilterable, gateway};
-use url::Url;
 
 use crate::{
-    client::key_manager::KeyManager,
     config::{persistence::key_pathfinder::ClientKeyPathfinder, Config, GatewayEndpointConfig},
     error::ClientCoreError,
+    init::helpers::{query_gateway_details, register_with_gateway_and_store_keys},
 };
+
+mod helpers;
 
 #[derive(Debug, Serialize)]
 pub struct InitResults {
@@ -62,84 +57,6 @@ impl Display for InitResults {
     }
 }
 
-pub async fn query_gateway_details(
-    validator_servers: Vec<Url>,
-    chosen_gateway_id: Option<&str>,
-) -> Result<gateway::Node, ClientCoreError> {
-    let validator_api = validator_servers
-        .choose(&mut thread_rng())
-        .ok_or(ClientCoreError::ListOfValidatorApisIsEmpty)?;
-    let validator_client = validator_client::client::ApiClient::new(validator_api.clone());
-
-    log::trace!("Fetching list of gateways from: {}", validator_api);
-    let gateways = validator_client.get_cached_gateways().await?;
-    let valid_gateways = gateways
-        .into_iter()
-        .filter_map(|gateway| gateway.try_into().ok())
-        .collect::<Vec<gateway::Node>>();
-
-    let filtered_gateways = valid_gateways.filter_by_version(env!("CARGO_PKG_VERSION"));
-
-    // if we have chosen particular gateway - use it, otherwise choose a random one.
-    // (remember that in active topology all gateways have at least 100 reputation so should
-    // be working correctly)
-    if let Some(gateway_id) = chosen_gateway_id {
-        filtered_gateways
-            .iter()
-            .find(|gateway| gateway.identity_key.to_base58_string() == gateway_id)
-            .ok_or_else(|| ClientCoreError::NoGatewayWithId(gateway_id.to_string()))
-            .cloned()
-    } else {
-        filtered_gateways
-            .choose(&mut rand::thread_rng())
-            .ok_or(ClientCoreError::NoGatewaysOnNetwork)
-            .cloned()
-    }
-}
-
-pub async fn register_with_gateway_and_store_keys<T>(
-    gateway_details: gateway::Node,
-    config: &Config<T>,
-) -> Result<(), ClientCoreError>
-where
-    T: NymConfig,
-{
-    let mut rng = OsRng;
-    let mut key_manager = KeyManager::new(&mut rng);
-
-    let shared_keys =
-        register_with_gateway(&gateway_details, key_manager.identity_keypair()).await?;
-    key_manager.insert_gateway_shared_key(shared_keys);
-
-    let pathfinder = ClientKeyPathfinder::new_from_config(config);
-    Ok(key_manager
-        .store_keys(&pathfinder)
-        .tap_err(|err| log::error!("Failed to generate keys: {err}"))?)
-}
-
-async fn register_with_gateway(
-    gateway: &gateway::Node,
-    our_identity: Arc<identity::KeyPair>,
-) -> Result<Arc<SharedKeys>, ClientCoreError> {
-    let timeout = Duration::from_millis(1500);
-    let mut gateway_client = GatewayClient::new_init(
-        gateway.clients_address(),
-        gateway.identity_key,
-        gateway.owner.clone(),
-        our_identity.clone(),
-        timeout,
-    );
-    gateway_client
-        .establish_connection()
-        .await
-        .tap_err(|_| log::warn!("Failed to establish connection with gateway!"))?;
-    let shared_keys = gateway_client
-        .perform_initial_authentication()
-        .await
-        .tap_err(|_| log::warn!("Failed to register with the gateway!"))?;
-    Ok(shared_keys)
-}
-
 pub async fn setup_gateway<T: NymConfig>(
     register: bool,
     user_chosen_gateway_id: Option<&str>,
@@ -175,7 +92,9 @@ pub async fn setup_gateway<T: NymConfig>(
     }
 }
 
-pub fn get_client_address<T>(config: &Config<T>) -> Result<Recipient, ClientCoreError>
+pub fn get_client_address_from_stored_keys<T>(
+    config: &Config<T>,
+) -> Result<Recipient, ClientCoreError>
 where
     T: config::NymConfig,
 {
