+++ clients/native/src/commands/init.rs
@@ -165,7 +165,7 @@ pub(crate) async fn execute(args: &Init) {
     );
     println!("Client configuration completed.");
 
-    let address = client_core::init::get_client_address(config.get_base()).unwrap_or_else(|err| {
+    let address = client_core::init::get_client_address_from_stored_keys(config.get_base()).unwrap_or_else(|err| {
         eprintln!("Failed to get address\nError: {err}");
         std::process::exit(1)
     });
